[New] added blueprints in flask in order to split main.py
