+++ odyssey/flask_validators/username_validator.py
@@ -1,0 +1,18 @@
+# Built-in imports
+import json
+
+# Third party library imports
+from flask import Blueprint, request, jsonify
+
+# Imports from .py files
+from flask_classes.user import User
+
+username_validator_bp = Blueprint('username_validator_bp', __name__)
+
+@username_validator_bp.route('/FpCerpd9Z7SIbjmN81Jy/username', methods=['POST'])
+def check_username():
+	username = request.get_json().get('username')
+	user = User.find_by_username(username)
+	if user:
+		return jsonify(success=False, message='An account with such username already exists!')
+	return jsonify(success=True)
